+++ user/views.py
@@ -8,6 +8,7 @@ from django.core.validators import RegexValidator
 from django.core.exceptions import ValidationError
 from django.views import View
 
+from .utils import login_decorator
 from trost.settings import SECRET_KEY, HASH_KEY
 from partner.models import (
         Counselor,
@@ -17,15 +18,11 @@ from .models import (
         History,
         User
 )
-from .models import User
-
 
 class SignUpView(View):
     def post(self, request):
         data=json.loads(request.body)
-        # regex of email address validation
         email_validator = RegexValidator(regex = '^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*.[a-zA-Z]{2,3}$')
-        # regex of password validation
         psw_validator = RegexValidator(regex='^.*(?=^.{6,14}$)(?=.*\d)(?=.*[a-zA-Z])(?=.*[!@#$%^&*?()_~]).*$')
 
         try:
@@ -36,7 +33,6 @@ class SignUpView(View):
             enc_pw=bcrypt.hashpw(data['password'].encode('utf-8'), bcrypt.gensalt())
             final_nickname = data['nickname']
             if not data['nickname']:
-                # random nickname generator
                 final_nickname = names.get_first_name()
             # saving data
             User(
@@ -44,10 +40,11 @@ class SignUpView(View):
                 password = enc_pw.decode('utf-8'),
                 nickname = final_nickname,
             ).save()
-        except ValidationError:
-            return HttpResponse(status= 400)
-        else:
             return JsonResponse({'message':'SUCCESS'}, status=200)
+        except ValidationError:
+            return JsonResponse({'message':'Validation error'}, status= 400)
+        except KeyError:
+            return JsonResponse({'message':'Key error'}, status= 400)
 
 # the button of checking email duplication
 class EmailCheckView(View):
@@ -87,20 +84,6 @@ class SignInView(View):
 
     def get(self, request):
         return HttpResponse(status=200)
-
-def login_decorator(func):
-    def wrapper(self, request, *args, **kwargs):
-        try:
-            token = request.headers.get('Authorization', None)
-            payload = jwt.decode(token, SECRET_KEY, algorithm=HASH_KEY)
-            user_mail = User.objects.get(email=payload['email'])
-            request.user = user_mail
-        except jwt.exceptions.DecodeError:
-            return JsonResponse({'message' : 'Invalid_token' }, status=400)
-        except Account.DoesNotExist:
-            return JsonResponse({'message' : 'Invalid_user'}, status=400)
-        return func(self, request, *args, **kwargs)
-    return wrapper
 
 class HistoryView(View):
     @login_decorator
